requirements.txt
anyio==4.13.0
caldav==3.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
fonttools==4.62.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icalendar==7.0.3
icalendar-searcher==1.0.5
idna==3.11
jh2==5.0.10
kiwisolver==1.5.0
lxml==6.0.2
matplotlib==3.10.1
niquests==3.18.2
numpy==2.4.4
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
PyYAML==6.0.3
qh3==1.7.0
recurring-ical-events==3.8.1
requests==2.33.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
urllib3-future==2.19.900
wassima==2.0.5
webdavclient3==3.14.6
x-wr-timezone==2.0.1
